Migrate to forks of st and dwm flexipatch
Before, I wasn't using a fork of the original projects and hence, I was
left behind on updates.
manually applying commits also felt like a chore.
therefore using a fork of the original projects to stay up to date with
the patches and changes much like dwmblocks-async felt like the right
thing to do.